Converted core/number tests to use Testify instead of Check.
